feat: rename utilities to hooks as the function is now a hook
